factor requirements into separate module
