+++ src/ecoscope-workflows-core/ecoscope_workflows_core/_models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+__all__ = ["_AllowArbitraryTypes"]
+
+
+class _AllowArbitraryTypes(BaseModel):
+    model_config = dict(arbitrary_types_allowed=True)
