+++ src/ecoscope-workflows-core/ecoscope_workflows_core/compiler.py
@@ -28,6 +28,7 @@ from pydantic.functional_validators import AfterValidator, BeforeValidator
 from ecoscope_workflows_core.artifacts import PixiToml
 from ecoscope_workflows_core.jsonschema import ReactJSONSchemaFormConfiguration
 from ecoscope_workflows_core.registry import KnownTask, known_tasks
+from ecoscope_workflows_core.requirements import CondaDependency
 
 T = TypeVar("T")
 
@@ -383,6 +384,13 @@ def ruff_formatted(returns_str_func: Callable[..., str]) -> Callable:
         return linted
 
     return wrapper
+
+
+class Requirements(_ForbidExtra):
+    """Requirements for a workflow."""
+
+    compile: list[CondaDependency]
+    run: list[CondaDependency]
 
 
 class Spec(_ForbidExtra):
